requirements.txt
numpy>=1.16.2
torch==1.1.0
torchgeometry>=0.1.2
PyYAML==5.1.1
configargparse
smplx
pillow
opencv-python
tqdm
human_body_prior
pyrender>=0.1.23
trimesh>=2.37.6
scipy
open3d-python==0.7.0.0
